+++ src/entities/event/iboundary_event_entity.ts
@@ -1,0 +1,6 @@
+import {INodeInstanceEntity} from '../index';
+import {IEventEntity} from './index';
+
+export interface IBoundaryEventEntity extends IEventEntity {
+  attachedToInstance: INodeInstanceEntity;
+}
