+++ src/entities/event/ievent_entity.ts
@@ -1,0 +1,5 @@
+import {INodeInstanceEntity} from '../index';
+
+export interface IEventEntity extends INodeInstanceEntity {
+  dispose(): void;
+}
