+++ src/entities/event/istart_event_entity.ts
@@ -1,0 +1,3 @@
+import {IEventEntity} from './index';
+
+export interface IStartEventEntity extends IEventEntity {}
