+++ src/entities/event/ithrow_event_entity.ts
@@ -1,0 +1,3 @@
+import {IEventEntity} from './index';
+
+export interface IThrowEventEntity extends IEventEntity {}
