+++ src/entities/gateway/iexclusive_gateway_entity.ts
@@ -1,0 +1,5 @@
+import {INodeInstanceEntity} from '../index';
+
+export interface IExclusiveGatewayEntity extends INodeInstanceEntity {
+  follow: any;
+}
