+++ src/entities/gateway/index.ts
@@ -1,0 +1,2 @@
+export * from './iexclusive_gateway_entity';
+export * from './iparallel_gateway_entity';
