+++ src/entities/gateway/iparallel_gateway_entity.ts
@@ -1,0 +1,5 @@
+import {INodeInstanceEntity} from '../index';
+
+export interface IParallelGatewayEntity extends INodeInstanceEntity {
+  parallelType: string;
+}
