+++ src/entities/process/index.ts
@@ -1,0 +1,7 @@
+export * from './iparam_start';
+export * from './iparam_update_defs';
+export * from './iprocess_def_entity';
+export * from './iprocess_entity';
+export * from './iprocess_token_entity';
+export * from './isubprocess_external_entity';
+export * from './isubprocess_internal_entity';
