+++ src/entities/process/iparam_start.ts
@@ -1,0 +1,9 @@
+export interface IParamStart {
+  id?: string;
+  key?: string;
+  initialToken?: any;
+  source?: any;
+  isSubProcess?: boolean;
+  participant?: string;
+  version?: string;
+}
