+++ src/entities/process/iparam_update_defs.ts
@@ -1,0 +1,5 @@
+import {IBpmnDiagram} from '../../index';
+
+export interface IParamUpdateDefs {
+  bpmnDiagram: IBpmnDiagram;
+}
