+++ src/entities/process/iprocess_token_entity.ts
@@ -1,0 +1,8 @@
+import {ExecutionContext, IEntity} from '@essential-projects/core_contracts';
+import {IProcessEntity} from './index';
+
+export interface IProcessTokenEntity extends IEntity {
+  data: any;
+  process: IProcessEntity;
+  getProcess(context: ExecutionContext): Promise<IProcessEntity>;
+}
