+++ src/entities/process/isubprocess_external_entity.ts
@@ -1,0 +1,3 @@
+import {INodeInstanceEntity} from '../index';
+
+export interface ISubprocessExternalEntity extends INodeInstanceEntity {}
