+++ src/entities/process/isubprocess_internal_entity.ts
@@ -1,0 +1,3 @@
+import {INodeInstanceEntity} from '../index';
+
+export interface ISubprocessInternalEntity extends INodeInstanceEntity {}
