+++ src/entities/task/index.ts
@@ -1,0 +1,3 @@
+export * from './iscript_task_entity';
+export * from './iservice_task_entity';
+export * from './iuser_task_entity';
