+++ src/entities/task/iscript_task_entity.ts
@@ -1,0 +1,5 @@
+import {INodeInstanceEntity} from '../index';
+
+export interface IScriptTaskEntity extends INodeInstanceEntity {
+  script: string;
+}
