+++ src/entities/task/iservice_task_entity.ts
@@ -1,0 +1,3 @@
+import {INodeInstanceEntity} from '../index';
+
+export interface IServiceTaskEntity extends INodeInstanceEntity {}
