+++ src/entities/task/iuser_task_entity.ts
@@ -1,0 +1,7 @@
+import {ExecutionContext} from '@essential-projects/core_contracts';
+import {IUserTaskMessageData} from '../../index';
+import {INodeInstanceEntity} from '../index';
+
+export interface IUserTaskEntity extends INodeInstanceEntity {
+  getUserTaskData(context: ExecutionContext): Promise<IUserTaskMessageData>;
+}
