+++ src/import/iimport_from_file_options.ts
@@ -1,0 +1,3 @@
+export interface IImportFromFileOptions {
+  overwriteExisting?: boolean;
+}
