+++ src/import/iimport_from_xml_options.ts
@@ -1,0 +1,3 @@
+import {IImportFromFileOptions} from './index';
+
+export interface IImportFromXmlOptions extends IImportFromFileOptions {}
