+++ src/import/iparam_import_from_file.ts
@@ -1,0 +1,3 @@
+export interface IParamImportFromFile {
+  file: string;
+}
