+++ src/import/iparam_import_from_xml.ts
@@ -1,0 +1,8 @@
+export interface IParamImportFromXml {
+  xml: string;
+  internalName?: string;
+  category?: string;
+  module?: string;
+  path?: string;
+  readonly?: boolean;
+}
