+++ src/index.ts
@@ -1,1 +1,12 @@
-export * from './interfaces';
+export * from './entities/index';
+export * from './import/index';
+
+export * from './constants';
+export * from './ibpmn_diagram';
+export * from './inode_instance_entity_type_service';
+export * from './iparams_continue_from_remote';
+export * from './iprocess_def_entity_type_service';
+export * from './iprocess_engine_service';
+export * from './iprocess_entry';
+export * from './iprocess_repository';
+export * from './iuser_task_message_data';
