+++ src/iparams_continue_from_remote.ts
@@ -1,0 +1,5 @@
+export interface IParamsContinueFromRemote {
+  nextDef: any;
+  source: any;
+  token: any;
+}
