+++ src/iprocess_engine_service.ts
@@ -1,0 +1,10 @@
+import {ExecutionContext, IPublicGetOptions} from '@essential-projects/core_contracts';
+import {IUserTaskMessageData} from './index';
+
+export interface IProcessEngineService {
+  config: any;
+  initialize(): Promise<void>;
+  start(context: ExecutionContext, data: any, options: IPublicGetOptions): Promise<string>;
+  getUserTaskData(context: ExecutionContext, userTaskId: string): Promise<IUserTaskMessageData>;
+  executeProcess(context: ExecutionContext, id: string, key: string, initialToken: any, version?: string): Promise<any>;
+}
