+++ src/iprocess_entry.ts
@@ -1,0 +1,8 @@
+export interface IProcessEntry {
+  name: string;
+  bpmnXml: string;
+  category: string;
+  module: string;
+  path: string;
+  readonly: boolean;
+}
