+++ src/iuser_task_message_data.ts
@@ -1,0 +1,8 @@
+import {IUserTaskEntity} from './entities/index';
+
+export interface IUserTaskMessageData {
+  userTaskEntity: IUserTaskEntity;
+  uiName: string;
+  uiData: any;
+  uiConfig: any;
+}
